+++ include/tag/byte_array_tag.h
@@ -132,7 +132,7 @@ public:
     /*
      * Returns a byte array tag value's size
      */
-    unsigned int size(void) { return value.size(); }
+    size_t size(void) { return value.size(); }
 
     /*
      * Return a string representation of a byte array tag
