+++ include/tag/int_array_tag.h
@@ -132,7 +132,7 @@ public:
     /*
      * Returns a integer array tag value's size
      */
-    unsigned int size(void) { return value.size(); }
+    size_t size(void) { return value.size(); }
 
     /*
      * Return a string representation of a integer array tag
